+++ lib/govuk_sidekiq/testing.rb
@@ -1,0 +1,5 @@
+require "sidekiq/testing"
+
+Sidekiq::Testing.server_middleware do |chain|
+  chain.add GovukSidekiq::APIHeaders::ServerMiddleware
+end